Extract out fixture code into its own module
Use the fixture feature flag now instead of the "fake" one for more clarity
