+++ iroh-ctl/src/main.rs
@@ -5,6 +5,6 @@ use clap::Parser;
 async fn main() -> Result<()> {
     let cli = iroh_ctl::run::Cli::parse();
     // run_cli exists in two versions, one for real client interaction,
-    // and one for testing purposes using a fake API implementation
+    // and one for testing purposes using mocked fixtures of the API
     iroh_ctl::run::run_cli(cli).await
 }
